chore: exported types from live market block
